requirements.txt
streamlit
pypdf
python-dotenv
chromadb
google-generativeai
pytesseract
Pillow
pdf2image
sentence-transformers
einops
